+++ .vscode/settings.json
@@ -1,4 +1,4 @@
 {
-  "cSpell.words": ["pubdev", "pubspec"],
+  "cSpell.words": ["lato", "pubdev", "pubspec"],
   "cmake.configureOnOpen": false
 }
