+++ lib/pages/home_detail.dart
@@ -1,0 +1,74 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_learning/widgets/theme.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+import '../models/catalog.dart';
+
+class HomeDetailPage extends StatelessWidget {
+  final Item catalog;
+
+  const HomeDetailPage({super.key, required this.catalog});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+      backgroundColor: MyTheme.creamColor,
+      bottomNavigationBar: Container(
+        color: Colors.white,
+        child: ButtonBar(
+          alignment: MainAxisAlignment.spaceBetween,
+          buttonPadding: EdgeInsets.zero,
+          children: [
+            "\$${catalog.price}".text.bold.xl4.red800.make(),
+            ElevatedButton(
+              onPressed: () {},
+              style: ButtonStyle(
+                  backgroundColor: MaterialStateProperty.all(
+                    MyTheme.darkBluishColor,
+                  ),
+                  shape: MaterialStateProperty.all(
+                    const StadiumBorder(),
+                  )),
+              child: "Buy".text.make(),
+            ).wh(100, 50)
+          ],
+        ).p32(),
+      ),
+      body: SafeArea(
+        bottom: false,
+        child: Column(
+          children: [
+            Hero(
+              tag: Key(
+                catalog.id.toString(),
+              ),
+              child: Image.network(catalog.image),
+            ).h32(context),
+            //.p32(),
+            Expanded(
+                child: VxArc(
+              height: 30.0,
+              arcType: VxArcType.CONVEY,
+              edge: VxEdge.TOP,
+              child: Container(
+                color: Colors.white,
+                width: context.screenWidth,
+                child: Column(
+                  children: [
+                    catalog.bookName.text.xl3
+                        .color(MyTheme.darkBluishColor)
+                        .bold
+                        .make(),
+                    catalog.desc.text.textStyle(context.captionStyle).xl.make(),
+                    10.heightBox
+                  ],
+                ).py64(),
+              ),
+            ))
+          ],
+        ),
+      ),
+    );
+  }
+}
