+++ lib/widgets/home_widgets/catalog_list.dart
@@ -1,0 +1,92 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_learning/pages/home_detail.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+import '../../models/catalog.dart';
+import '../theme.dart';
+import 'catalog_image.dart';
+
+class CatalogList extends StatelessWidget {
+  const CatalogList({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return ListView.builder(
+      shrinkWrap: true, //
+      itemCount: CatalogModel.items.length,
+      itemBuilder: (context, index) {
+        final catalog = CatalogModel.items[index];
+        return InkWell(
+          onTap: () => Navigator.push(
+            context,
+            MaterialPageRoute(
+              builder: (context) => HomeDetailPage(catalog: catalog),
+            ),
+          ),
+          child: CatalogItem(
+            catalog: catalog,
+            key: const Key(""),
+          ),
+        );
+      },
+    );
+  }
+}
+
+class CatalogItem extends StatelessWidget {
+  final Item catalog;
+
+  const CatalogItem({required Key key, required this.catalog})
+      : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return VxBox(
+      //like container
+      child: Row(
+        children: [
+          Hero(
+            tag: Key(
+              catalog.id.toString(),
+            ),
+            child: CatalogImage(
+              image: catalog.image,
+              key: const Key(""),
+            ),
+          ),
+          Expanded(
+              child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              catalog.bookName.text.lg
+                  .color(MyTheme.darkBluishColor)
+                  .bold
+                  .make(),
+              catalog.desc.text.textStyle(context.captionStyle).make(),
+              10.heightBox,
+              ButtonBar(
+                alignment: MainAxisAlignment.spaceBetween,
+                buttonPadding: EdgeInsets.zero,
+                children: [
+                  "\$${catalog.price}".text.bold.xl.make(),
+                  ElevatedButton(
+                    onPressed: () {},
+                    style: ButtonStyle(
+                        backgroundColor: MaterialStateProperty.all(
+                          MyTheme.darkBluishColor,
+                        ),
+                        shape: MaterialStateProperty.all(
+                          const StadiumBorder(),
+                        )),
+                    child: "Buy".text.make(),
+                  )
+                ],
+              ).pOnly(right: 8.0)
+            ],
+          ))
+        ],
+      ),
+    ).white.rounded.square(150).make().py16();
+  }
+}
